// projects/mss-invoice/src/app/core/models/company.model.ts
export interface Company {
  id?: number;
  legalName: string;
  legalNameAr?: string;
  addressLine?: string;
  poBox?: string;
  postalCode?: string;
  city?: string;
  country?: string;
  phone?: string;
  email?: string;
  website?: string;
  vatNumber?: string;
  taxCardNumber?: string;
  mofCode?: string;
  bankName?: string;
  bankAccountName?: string;
  bankAccountNumber?: string;
  bankIban?: string;
  bankSwift?: string;
  invoiceFooterNote?: string;
  logoPath?: string;
}